refactor(controller, instrumenter): extract instrumentation logic
Extract the controller's instrumentation logic into a separate module,
called instrumentation/instrumenter. This allows running the
InstrumentAtStartup task with a separate client and to move it into
executeStartupTasks; getting rid of the arbitrary (and potentially
problematic) 10 second wait, which we previously needed to wait until
the manager had finished initializing its Kubernetes API client.
