+++ internal/backendconnection/otelcolresources/otelcol_resources_suite_test.go
@@ -1,3 +1,6 @@
+// SPDX-FileCopyrightText: Copyright 2024 Dash0 Inc.
+// SPDX-License-Identifier: Apache-2.0
+
 package otelcolresources
 
 import (
