+++ internal/dash0/instrumentation/instrumentable_workloads.go
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: Copyright 2024 Dash0 Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-package controller
+package instrumentation
 
 import (
 	"github.com/go-logr/logr"
