+++ internal/dash0/webhook/dash0_webhook_test.go
@@ -199,7 +199,7 @@ var _ = Describe("The Dash0 webhook", func() {
 							EnvVars:                                  3,
 							NodeOptionsEnvVarIdx:                     1,
 							Dash0CollectorBaseUrlEnvVarIdx:           2,
-							Dash0CollectorBaseUrlEnvVarExpectedValue: "http://dash0-operator-opentelemetry-collector.dash0-system.svc.cluster.local:4318",
+							Dash0CollectorBaseUrlEnvVarExpectedValue: OTelCollectorBaseUrlTest,
 						},
 						{
 							VolumeMounts:                             3,
@@ -207,7 +207,7 @@ var _ = Describe("The Dash0 webhook", func() {
 							EnvVars:                                  4,
 							NodeOptionsEnvVarIdx:                     2,
 							Dash0CollectorBaseUrlEnvVarIdx:           3,
-							Dash0CollectorBaseUrlEnvVarExpectedValue: "http://dash0-operator-opentelemetry-collector.dash0-system.svc.cluster.local:4318",
+							Dash0CollectorBaseUrlEnvVarExpectedValue: OTelCollectorBaseUrlTest,
 						},
 					},
 				})
@@ -236,7 +236,7 @@ var _ = Describe("The Dash0 webhook", func() {
 							NodeOptionsEnvVarIdx:                     1,
 							NodeOptionsUsesValueFrom:                 true,
 							Dash0CollectorBaseUrlEnvVarIdx:           2,
-							Dash0CollectorBaseUrlEnvVarExpectedValue: "http://dash0-operator-opentelemetry-collector.dash0-system.svc.cluster.local:4318",
+							Dash0CollectorBaseUrlEnvVarExpectedValue: OTelCollectorBaseUrlTest,
 						},
 						{
 							VolumeMounts:                             3,
@@ -245,7 +245,7 @@ var _ = Describe("The Dash0 webhook", func() {
 							NodeOptionsEnvVarIdx:                     1,
 							NodeOptionsValue:                         "--require /__dash0__/instrumentation/node.js/node_modules/@dash0hq/opentelemetry --require something-else --experimental-modules",
 							Dash0CollectorBaseUrlEnvVarIdx:           0,
-							Dash0CollectorBaseUrlEnvVarExpectedValue: "http://dash0-operator-opentelemetry-collector.dash0-system.svc.cluster.local:4318",
+							Dash0CollectorBaseUrlEnvVarExpectedValue: OTelCollectorBaseUrlTest,
 						},
 					},
 				})
