+++ internal/dash0/workloads/workloads_suite_test.go
@@ -1,4 +1,7 @@
-package workloads_test
+// SPDX-FileCopyrightText: Copyright 2024 Dash0 Inc.
+// SPDX-License-Identifier: Apache-2.0
+
+package workloads
 
 import (
 	"testing"
