+++ test/util/constants.go
@@ -28,10 +28,11 @@ const (
 	ConfigurationReloaderImageTest = "some-registry.com:1234/dash0hq/configuration-reloader:10.11.12"
 	FilelogOffsetSynchImageTest    = "some-registry.com:1234/dash0hq/filelog-offset-synch:13.14.15"
 
-	IngressEndpointTest    = "ingress.endpoint.dash0.com:4317"
-	AuthorizationTokenTest = "authorization-token"
-	SecretRefTest          = "secret-ref"
-	SecretRefEmpty         = ""
+	OTelCollectorBaseUrlTest = "http://dash0-operator-opentelemetry-collector.dash0-system.svc.cluster.local:4318"
+	IngressEndpointTest      = "ingress.endpoint.dash0.com:4317"
+	AuthorizationTokenTest   = "authorization-token"
+	SecretRefTest            = "secret-ref"
+	SecretRefEmpty           = ""
 )
 
 var (
