+++ test/util/resources.go
@@ -819,7 +819,7 @@ func InstrumentedDeploymentWithMoreBellsAndWhistles(namespace string, name strin
 				},
 				{
 					Name:  "DASH0_OTEL_COLLECTOR_BASE_URL",
-					Value: "http://dash0-operator-opentelemetry-collector.dash0-system.svc.cluster.local:4318",
+					Value: OTelCollectorBaseUrlTest,
 				},
 			},
 		},
@@ -855,7 +855,7 @@ func InstrumentedDeploymentWithMoreBellsAndWhistles(namespace string, name strin
 				},
 				{
 					Name:  "DASH0_OTEL_COLLECTOR_BASE_URL",
-					Value: "http://dash0-operator-opentelemetry-collector.dash0-system.svc.cluster.local:4318",
+					Value: OTelCollectorBaseUrlTest,
 				},
 			},
 		},
@@ -894,7 +894,7 @@ func simulateInstrumentedPodSpec(podSpec *corev1.PodSpec, meta *metav1.ObjectMet
 		},
 		{
 			Name:  "DASH0_OTEL_COLLECTOR_BASE_URL",
-			Value: "http://dash0-operator-opentelemetry-collector.dash0-system.svc.cluster.local:4318",
+			Value: OTelCollectorBaseUrlTest,
 		},
 	}
